Revert "OP-3682 - extracted sha256 method to lib"
This reverts commit c4854be5
